* sensors nested under experiments * messaging and reactions nested under experiments TODO: speed in Hz
