+++ composeApp/src/commonMain/kotlin/data/shared/GeneralObserver.kt
@@ -6,20 +6,24 @@ import ui.conversation.ConversationState
 
 sealed class GeneralObserver<D> {
     abstract fun invoke(data: D)
+    abstract val key: String?
 
     class MessageObserver(
+        override val key: String?,
         private val invocation: (FullConversationMessage) -> Unit
     ): GeneralObserver<FullConversationMessage>() {
         override fun invoke(data: FullConversationMessage) = invocation(data)
     }
 
     class ReactionsObserver(
+        override val key: String?,
         private val invocation: (MessageReactionIO) -> Unit
     ): GeneralObserver<MessageReactionIO>() {
         override fun invoke(data: MessageReactionIO) = invocation(data)
     }
 
     class ConversationStateObserver(
+        override val key: String?,
         private val invocation: (ConversationState) -> Unit
     ): GeneralObserver<ConversationState>() {
         override fun invoke(data: ConversationState) = invocation(data)
