+++ composeApp/src/commonMain/kotlin/ui/dev/experiment/ExperimentRepository.kt
@@ -4,11 +4,12 @@ import data.io.experiment.ExperimentIO
 import data.io.experiment.ExperimentSet
 import database.dao.experiment.ExperimentDao
 import database.dao.experiment.ExperimentSetDao
+import ui.dev.DeveloperConsoleRepository
 
 class ExperimentRepository(
     private val experimentDao: ExperimentDao,
     private val experimentSetDao: ExperimentSetDao
-) {
+): DeveloperConsoleRepository() {
     suspend fun insertSet(set: ExperimentSet) = experimentSetDao.insert(set)
     suspend fun removeSet(uid: String) = experimentSetDao.remove(uid)
     suspend fun getExperiments(owner: String?) = experimentDao.getAll(owner)
