+++ composeApp/src/commonMain/kotlin/ui/dev/experiment/ExperimentsContent.kt
@@ -2,9 +2,7 @@ package ui.dev.experiment
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.Crossfade
-import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.background
-import androidx.compose.foundation.focusable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -12,65 +10,56 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.requiredHeight
 import androidx.compose.foundation.layout.requiredSize
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.foundation.text.input.InputTransformation
-import androidx.compose.foundation.text.input.TextFieldLineLimits
 import androidx.compose.foundation.text.input.TextFieldState
-import androidx.compose.foundation.text.input.byValue
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.outlined.Add
-import androidx.compose.material.icons.outlined.CheckBox
-import androidx.compose.material.icons.outlined.Delete
-import androidx.compose.material.icons.outlined.EventRepeat
-import androidx.compose.material.icons.outlined.PlayArrow
-import androidx.compose.material.icons.outlined.Quiz
+import androidx.compose.material.icons.outlined.Check
+import androidx.compose.material.icons.outlined.Folder
 import androidx.compose.material.icons.outlined.Stop
-import androidx.compose.material.icons.outlined.Tag
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import augmy.interactive.shared.ext.scalingClickable
 import augmy.interactive.shared.ui.components.ComponentHeaderButton
-import augmy.interactive.shared.ui.components.ProgressPressableContainer
-import augmy.interactive.shared.ui.components.SimpleModalBottomSheet
-import augmy.interactive.shared.ui.components.dialog.AlertDialog
-import augmy.interactive.shared.ui.components.dialog.ButtonState
+import augmy.interactive.shared.ui.components.ContrastHeaderButton
+import augmy.interactive.shared.ui.components.LoadingHeaderButton
+import augmy.interactive.shared.ui.components.MultiChoiceSwitchMinimalistic
 import augmy.interactive.shared.ui.components.input.CustomTextField
-import augmy.interactive.shared.ui.components.input.DELAY_BETWEEN_TYPING_SHORT
+import augmy.interactive.shared.ui.components.rememberMultiChoiceState
 import augmy.interactive.shared.ui.theme.LocalTheme
 import augmy.interactive.shared.ui.theme.SharedColors
+import data.io.base.BaseResponse
 import data.io.experiment.ExperimentIO
 import data.io.experiment.FullExperiment
-import kotlinx.coroutines.cancelChildren
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
+import data.sensor.SensorDelay
+import io.github.vinceglb.filekit.dialogs.compose.rememberFileSaverLauncher
 import org.koin.compose.viewmodel.koinViewModel
-import kotlin.uuid.ExperimentalUuidApi
+import ui.dev.DeveloperConsoleModel
 
 @Composable
-fun ExperimentContent(model: ExperimentModel = koinViewModel()) {
+fun ExperimentContent(
+    developerModel: DeveloperConsoleModel,
+    model: ExperimentModel = koinViewModel()
+) {
     val experiments = model.experiments.collectAsState()
     val activeExperiments = model.activeExperiments.collectAsState()
 
@@ -80,15 +69,25 @@ fun ExperimentContent(model: ExperimentModel = koinViewModel()) {
 
     selectedExperiment.value?.let { experiment ->
         ExperimentBottomSheet(
-            model = model,
             experiment = experiment,
             onDismissRequest = { selectedExperiment.value = null }
         )
     }
 
+    LaunchedEffect(experiments.value) {
+        if (selectedExperiment.value != null) {
+            selectedExperiment.value = experiments.value.find {
+                it.data.uid == selectedExperiment.value?.data?.uid
+            }
+        }
+    }
+
     LazyColumn(
         verticalArrangement = Arrangement.spacedBy(6.dp)
     ) {
+        item(key = "streamingSection") {
+            StreamingSection(model = developerModel)
+        }
         item(key = "title") {
             Text(
                 modifier = Modifier
@@ -126,7 +125,10 @@ fun ExperimentContent(model: ExperimentModel = koinViewModel()) {
             ) {
                 Row(
                     modifier = Modifier
-                        .scalingClickable(key = experiment.data.uid, scaleInto = .95f) {
+                        .scalingClickable(
+                            key = experiment.data.toString(),
+                            scaleInto = .95f
+                        ) {
                             selectedExperiment.value = experiment
                         }
                         .background(
@@ -198,440 +200,183 @@ fun ExperimentContent(model: ExperimentModel = koinViewModel()) {
     }
 }
 
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalUuidApi::class)
 @Composable
-private fun ExperimentBottomSheet(
-    model: ExperimentModel,
-    experiment: FullExperiment,
-    onDismissRequest: () -> Unit
-) {
-    val cancellableScope = rememberCoroutineScope()
-    val activeExperiments = model.activeExperiments.collectAsState()
+private fun StreamingSection(model: DeveloperConsoleModel) {
+    val streamingUrlResponse = model.streamingUrlResponse.collectAsState()
 
-    val showDeleteConfirmation = remember(experiment.data.uid) {
-        mutableStateOf(false)
-    }
-    val showSetSelection = remember(experiment.data.uid) {
-        mutableStateOf(false)
-    }
-    val showFrequencySelection = remember(experiment.data.uid) {
-        mutableStateOf(false)
-    }
-    val showBehaviorSelection = remember(experiment.data.uid) {
-        mutableStateOf(false)
-    }
-    val selectedSet = remember(experiment.data.uid) {
-        mutableStateOf(experiment.sets.firstOrNull())
-    }
-    val selectedFrequency = remember(experiment.data.uid) {
-        mutableStateOf(experiment.data.displayFrequency)
-    }
-    val selectedBehavior = remember(experiment.data.uid) {
-        mutableStateOf(experiment.data.choiceBehavior)
-    }
-    val nameState = remember(experiment.data.uid) {
-        TextFieldState(initialText = experiment.data.name)
+    val streamingUrlState = remember(model) {
+        TextFieldState(initialText = model.streamingUrl)
     }
 
-    LaunchedEffect(nameState.text) {
-        cancellableScope.coroutineContext.cancelChildren()
-        cancellableScope.launch {
-            delay(DELAY_BETWEEN_TYPING_SHORT)
-            model.changeNameOf(experiment.data.uid, nameState.text)
+    val filePicker = rememberFileSaverLauncher(
+        onResult = { filePicker ->
+            model.setUpLocalStream(filePicker)
         }
-    }
+    )
 
-    when {
-        showDeleteConfirmation.value -> {
-            AlertDialog(
-                title = "Delete experiment",
-                message = AnnotatedString("Are you sure you want to delete this experiment?"),
-                dismissButtonState = ButtonState("Dismiss"),
-                confirmButtonState = ButtonState(
-                    text = "Confirm",
-                    onClick = {
-                        model.deleteExperiment(experiment.data.uid)
-                        showDeleteConfirmation.value = false
-                        onDismissRequest()
-                    }
-                ),
-                onDismissRequest = { showDeleteConfirmation.value = false }
-            )
-        }
-        showSetSelection.value -> {
-            SetListDialog(
-                model = model,
-                experiment = experiment,
-                selectedSet = selectedSet.value,
-                onConfirm = {
-                    selectedSet.value = it
-                    model.changeSetOf(experiment.data.uid, it.uid)
-                    showSetSelection.value = false
-                },
-                onDismissRequest = { showSetSelection.value = false }
-            )
-        }
-        showFrequencySelection.value -> {
-            DisplayFrequencyDialog(
-                initialFrequency = selectedFrequency.value,
-                onConfirm = {
-                    model.changeFrequencyOf(experiment.data.uid, it)
-                    selectedFrequency.value = it
-                },
-                onDismissRequest = { showFrequencySelection.value = false }
-            )
-        }
-        showBehaviorSelection.value -> ChoiceBehaviorDialog(
-            initialBehavior = selectedBehavior.value,
-            onConfirm = {
-                model.changeBehaviorOf(experiment.data.uid, it)
-                selectedBehavior.value = it
-            },
-            onDismissRequest = { showBehaviorSelection.value = false }
+    Text(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(color = LocalTheme.current.colors.appbarBackground)
+            .padding(horizontal = 12.dp, vertical = 6.dp),
+        text = "Data streaming",
+        style = LocalTheme.current.styles.subheading.copy(
+            color = LocalTheme.current.colors.appbarContent
         )
-    }
+    )
 
-    SimpleModalBottomSheet(
-        onDismissRequest = onDismissRequest
+    Row(
+        modifier = Modifier.padding(top = 8.dp),
+        verticalAlignment = Alignment.CenterVertically
     ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically
+        AnimatedVisibility(
+            modifier = Modifier.weight(1f),
+            visible = streamingUrlResponse.value !is BaseResponse.Success
         ) {
             CustomTextField(
-                modifier = Modifier.fillMaxWidth(.7f),
-                state = nameState,
-                prefixIcon = Icons.Outlined.Tag
+                keyboardOptions = KeyboardOptions(
+                    keyboardType = KeyboardType.Uri,
+                    imeAction = ImeAction.Done
+                ),
+                onKeyboardAction = {
+                    model.setupRemoteStream(streamingUrlState.text)
+                },
+                backgroundColor = LocalTheme.current.colors.backgroundLight,
+                hint = "Server url",
+                state = streamingUrlState,
+                shape = LocalTheme.current.shapes.componentShape,
+                errorText = (streamingUrlResponse.value as? BaseResponse.Error)?.message
             )
-
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(6.dp)
-            ) {
-                ProgressPressableContainer(
-                    modifier = Modifier.requiredSize(36.dp),
-                    onFinish = {
-                        showDeleteConfirmation.value = true
-                    },
-                    trackColor = LocalTheme.current.colors.disabled,
-                    progressColor = SharedColors.RED_ERROR
-                ) {
-                    Icon(
-                        modifier = Modifier.size(32.dp),
-                        imageVector = Icons.Outlined.Delete,
-                        contentDescription = null,
-                        tint = SharedColors.RED_ERROR
+        }
+        Crossfade(targetState = streamingUrlResponse.value) { response ->
+            when(response) {
+                is BaseResponse.Success -> {
+                    ContrastHeaderButton(
+                        text = "Stop remote stream",
+                        endImageVector = Icons.Outlined.Stop,
+                        contentColor = Color.White,
+                        containerColor = SharedColors.RED_ERROR,
+                        onClick = {
+                            model.stopRemoteStream()
+                        }
                     )
                 }
-
-                Crossfade(activeExperiments.value.contains(experiment.data.uid)) { isActive ->
-                    Icon(
-                        modifier = Modifier
-                            .scalingClickable {
-                                model.toggleExperiment(experiment.data.uid, !isActive)
-                                onDismissRequest()
-                            }
-                            .size(42.dp)
-                            .padding(5.dp),
-                        imageVector = if (isActive) Icons.Outlined.Stop else Icons.Outlined.PlayArrow,
-                        contentDescription = null,
-                        tint = if (isActive) SharedColors.RED_ERROR else LocalTheme.current.colors.brandMain
+                else -> {
+                    LoadingHeaderButton(
+                        text = "Stream",
+                        isLoading = streamingUrlResponse.value is BaseResponse.Loading,
+                        isEnabled = streamingUrlResponse.value !is BaseResponse.Loading,
+                        endImageVector = Icons.Outlined.Check,
+                        onClick = {
+                            model.setupRemoteStream(streamingUrlState.text)
+                        }
                     )
                 }
             }
         }
+        AnimatedVisibility(streamingUrlResponse.value is BaseResponse.Success) {
+            val selectedDelayIndex = rememberSaveable {
+                mutableStateOf(model.remoteStreamDelay.ordinal)
+            }
 
-        // value set
-        Row(
-            modifier = Modifier.padding(top = 12.dp, start = 16.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                modifier = Modifier.size(24.dp),
-                imageVector = Icons.Outlined.Quiz,
-                tint = LocalTheme.current.colors.disabled,
-                contentDescription = null
-            )
-            Text(
-                modifier = Modifier.padding(start = 2.dp),
-                text = "Values set: ",
-                style = LocalTheme.current.styles.regular
-            )
-            Text(
-                modifier = Modifier
-                    .padding(start = 6.dp)
-                    .scalingClickable {
-                        showSetSelection.value = true
-                    }
-                    .background(
-                        color = LocalTheme.current.colors.backgroundLight,
-                        shape = LocalTheme.current.shapes.rectangularActionShape
-                    )
-                    .padding(horizontal = 6.dp, vertical = 4.dp),
-                text = selectedSet.value?.name ?: "no set selected",
-                style = LocalTheme.current.styles.category
-            )
-        }
-
-        // frequency
-        Row(
-            modifier = Modifier.padding(top = 6.dp, start = 16.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                modifier = Modifier.size(24.dp),
-                imageVector = Icons.Outlined.EventRepeat,
-                tint = LocalTheme.current.colors.disabled,
-                contentDescription = null
-            )
-            Text(
-                modifier = Modifier.padding(start = 4.dp),
-                text = "Display frequency: ",
-                style = LocalTheme.current.styles.regular
-            )
-            Text(
-                modifier = Modifier
-                    .padding(start = 6.dp)
-                    .scalingClickable {
-                        showFrequencySelection.value = true
-                    }
-                    .background(
-                        color = LocalTheme.current.colors.backgroundLight,
-                        shape = LocalTheme.current.shapes.rectangularActionShape
-                    )
-                    .padding(horizontal = 6.dp, vertical = 4.dp),
-                text = when (val frequency = selectedFrequency.value) {
-                    is ExperimentIO.DisplayFrequency.Constant -> "every ${frequency.delaySeconds}s"
-                    ExperimentIO.DisplayFrequency.BeginEnd -> "start + end"
-                    ExperimentIO.DisplayFrequency.Permanent -> "all the time"
-                },
-                style = LocalTheme.current.styles.category
-            )
-        }
-
-        // choice behavior
-        Row(
-            modifier = Modifier.padding(top = 6.dp, start = 16.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                modifier = Modifier.size(24.dp),
-                imageVector = Icons.Outlined.CheckBox,
-                tint = LocalTheme.current.colors.disabled,
-                contentDescription = null
-            )
-            Text(
-                modifier = Modifier.padding(start = 4.dp),
-                text = "Choice behavior: ",
-                style = LocalTheme.current.styles.regular
-            )
-            Text(
-                modifier = Modifier
-                    .padding(start = 6.dp)
-                    .scalingClickable {
-                        showBehaviorSelection.value = true
-                    }
-                    .background(
-                        color = LocalTheme.current.colors.backgroundLight,
-                        shape = LocalTheme.current.shapes.rectangularActionShape
-                    )
-                    .padding(horizontal = 6.dp, vertical = 4.dp),
-                text = when (selectedBehavior.value) {
-                    ExperimentIO.ChoiceBehavior.SingleChoice -> "single-choice"
-                    ExperimentIO.ChoiceBehavior.MultiChoice -> "multi-choice"
-                    ExperimentIO.ChoiceBehavior.OrderedChoice -> "ordered-choice"
-                },
-                style = LocalTheme.current.styles.category
-            )
-        }
-
-
-        Spacer(Modifier.height(48.dp))
-    }
-}
-
-@Composable
-private fun DisplayFrequencyDialog(
-    initialFrequency: ExperimentIO.DisplayFrequency,
-    onConfirm: (ExperimentIO.DisplayFrequency) -> Unit,
-    onDismissRequest: () -> Unit
-) {
-    val missFocusRequester = remember { FocusRequester() }
-    val selectedFrequency = remember(initialFrequency) {
-        mutableStateOf(initialFrequency)
-    }
-
-    AlertDialog(
-        modifier = Modifier
-            .focusRequester(missFocusRequester)
-            .focusable()
-            .animateContentSize(),
-        title = "Select display frequency",
-        intrinsicContent = false,
-        additionalContent = {
             Row(
-                modifier = Modifier.scalingClickable(hoverEnabled = false, scaleInto = .95f) {
-                    selectedFrequency.value = ExperimentIO.DisplayFrequency.BeginEnd
-                    missFocusRequester.requestFocus()
-                },
+                modifier = Modifier.padding(start = 6.dp),
                 verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
+                horizontalArrangement = Arrangement.SpaceBetween
             ) {
-                RadioButton(
-                    colors = LocalTheme.current.styles.radioButtonColors,
-                    selected = selectedFrequency.value == ExperimentIO.DisplayFrequency.BeginEnd,
-                    onClick = {
-                        selectedFrequency.value = ExperimentIO.DisplayFrequency.BeginEnd
-                        missFocusRequester.requestFocus()
-                    }
-                )
                 Text(
-                    text = "Start + end",
+                    text = "Upload by: ",
                     style = LocalTheme.current.styles.regular
                 )
-            }
-            Row(
-                modifier = Modifier.scalingClickable(hoverEnabled = false, scaleInto = .95f) {
-                    selectedFrequency.value = ExperimentIO.DisplayFrequency.Permanent
-                    missFocusRequester.requestFocus()
-                },
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                RadioButton(
-                    colors = LocalTheme.current.styles.radioButtonColors,
-                    selected = selectedFrequency.value == ExperimentIO.DisplayFrequency.Permanent,
-                    onClick = {
-                        selectedFrequency.value = ExperimentIO.DisplayFrequency.Permanent
-                        missFocusRequester.requestFocus()
-                    }
-                )
-                Text(
-                    text = "All the time",
-                    style = LocalTheme.current.styles.regular
-                )
-            }
-            Row(
-                modifier = Modifier.scalingClickable(hoverEnabled = false, scaleInto = .95f) {
-                    selectedFrequency.value = ExperimentIO.DisplayFrequency.Constant(5)
-                },
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                val focusRequester = remember { FocusRequester() }
-                val isFocused = rememberSaveable(initialFrequency) {
-                    mutableStateOf(false)
-                }
-                val secondsState = remember(initialFrequency) {
-                    TextFieldState(
-                        initialText = (initialFrequency as? ExperimentIO.DisplayFrequency.Constant)?.delaySeconds?.toString() ?: "5"
-                    )
-                }
-
-                LaunchedEffect(isFocused.value) {
-                    if (isFocused.value) {
-                        selectedFrequency.value = ExperimentIO.DisplayFrequency.Constant(
-                            secondsState.text.toString().toLongOrNull() ?: 5L
-                        )
-                    }
-                }
-                LaunchedEffect(secondsState.text) {
-                    if (selectedFrequency.value is ExperimentIO.DisplayFrequency.Constant) {
-                        selectedFrequency.value = ExperimentIO.DisplayFrequency.Constant(
-                            secondsState.text.toString().toLongOrNull() ?: 5L
-                        )
-                    }
-                }
-
-                RadioButton(
-                    colors = LocalTheme.current.styles.radioButtonColors,
-                    selected = selectedFrequency.value is ExperimentIO.DisplayFrequency.Constant,
-                    onClick = {
-                        focusRequester.requestFocus()
-                    }
-                )
-                CustomTextField(
-                    backgroundColor = LocalTheme.current.colors.backgroundLight,
-                    state = secondsState,
-                    lineLimits = TextFieldLineLimits.SingleLine,
-                    hint = "Every X",
-                    keyboardOptions = KeyboardOptions(
-                        keyboardType = KeyboardType.Number,
-                        imeAction = ImeAction.Done
+                MultiChoiceSwitchMinimalistic(
+                    modifier = Modifier.padding(start = 6.dp),
+                    state = rememberMultiChoiceState(
+                        selectedTabIndex = selectedDelayIndex,
+                        items = SensorDelay.entries.map { it.name }.toMutableList()
                     ),
-                    focusRequester = focusRequester,
-                    inputTransformation = InputTransformation.byValue { _, proposed ->
-                        proposed.replace(Regex("[^0-9]"), "")
+                    onClick = { index ->
+                        selectedDelayIndex.value = index
+                        model.remoteStreamDelay =  SensorDelay.entries[index]
                     },
-                    isFocused = isFocused,
-                    trailingIcon = {
+                    onItemCreation = { _, index, _ ->
                         Text(
-                            text = "seconds",
-                            style = LocalTheme.current.styles.regular.copy(
-                                color = LocalTheme.current.colors.disabled
+                            modifier = Modifier
+                                .fillMaxWidth()
+                                .padding(horizontal = 6.dp),
+                            text = when(SensorDelay.entries[index]) {
+                                SensorDelay.Slow -> 50
+                                SensorDelay.Normal -> 20
+                                SensorDelay.Fast -> 1
+                            }.toString(),
+                            style = LocalTheme.current.styles.category.copy(
+                                textAlign = TextAlign.Center
                             )
                         )
                     }
                 )
             }
-        },
-        confirmButtonState = ButtonState(
-            text = "Confirm",
-            onClick = {
-                onConfirm(selectedFrequency.value)
-            }
-        ),
-        onDismissRequest = onDismissRequest
-    )
-}
-
-@Composable
-private fun ChoiceBehaviorDialog(
-    initialBehavior: ExperimentIO.ChoiceBehavior,
-    onConfirm: (ExperimentIO.ChoiceBehavior,) -> Unit,
-    onDismissRequest: () -> Unit
-) {
-    val selectedBehavior = remember(initialBehavior) {
-        mutableStateOf(initialBehavior)
+        }
     }
 
-    AlertDialog(
-        title = "Select choice behavior",
-        intrinsicContent = false,
-        additionalContent = {
-            ExperimentIO.ChoiceBehavior.entries.forEach { behavior ->
-                Row(
-                    modifier = Modifier.scalingClickable(hoverEnabled = false, scaleInto = .95f) {
-                        selectedBehavior.value = behavior
-                    },
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    RadioButton(
-                        colors = LocalTheme.current.styles.radioButtonColors,
-                        selected = selectedBehavior.value == behavior,
-                        onClick = {
-                            selectedBehavior.value = behavior
-                        }
-                    )
-                    Text(
-                        text = when (behavior) {
-                            ExperimentIO.ChoiceBehavior.SingleChoice -> "single-choice"
-                            ExperimentIO.ChoiceBehavior.MultiChoice -> "multi-choice"
-                            ExperimentIO.ChoiceBehavior.OrderedChoice -> "ordered-choice"
-                        },
-                        style = LocalTheme.current.styles.regular
-                    )
+    val localRunning = model.isLocalStreamRunning.collectAsState()
+    Crossfade(
+        modifier = Modifier.padding(top = 8.dp),
+        targetState = localRunning.value
+    ) {
+        if (it) {
+            ContrastHeaderButton(
+                text = "Stop local stream",
+                endImageVector = Icons.Outlined.Stop,
+                contentColor = Color.White,
+                containerColor = SharedColors.RED_ERROR,
+                onClick = {
+                    model.stopLocalStream()
                 }
+            )
+        }else {
+            ContrastHeaderButton(
+                text = "Stream locally",
+                endImageVector = Icons.Outlined.Folder,
+                contentColor = LocalTheme.current.colors.tetrial,
+                containerColor = LocalTheme.current.colors.brandMainDark,
+                onClick = {
+                    filePicker.launch(extension = "txt", suggestedName = "stream-sensory-augmy")
+                }
+            )
+        }
+    }
+
+    val streamLines = model.streamLines.collectAsState()
+    AnimatedVisibility(
+        modifier = Modifier
+            .padding(top = 8.dp)
+            .fillMaxWidth(),
+        visible = streamLines.value.isNotEmpty()
+    ) {
+        val state = rememberLazyListState()
+
+        LaunchedEffect(streamLines.value.size) {
+            state.animateScrollToItem(0)
+        }
+
+        LazyColumn(
+            modifier = Modifier
+                .requiredHeight(50.dp)
+                .padding(horizontal = 8.dp),
+            state = state,
+            verticalArrangement = Arrangement.spacedBy(6.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            reverseLayout = true
+        ) {
+            items(items = streamLines.value) { line ->
+                Text(
+                    modifier = Modifier.animateItem(),
+                    text = line,
+                    style = LocalTheme.current.styles.regular.copy(
+                        color = LocalTheme.current.colors.disabled
+                    )
+                )
             }
-        },
-        confirmButtonState = ButtonState(
-            text = "Confirm",
-            onClick = {
-                onConfirm(selectedBehavior.value)
-            }
-        ),
-        onDismissRequest = onDismissRequest
-    )
+        }
+    }
 }
